backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# AI & Embeddings
chromadb==0.4.22
openai==1.10.0
google-generativeai==0.3.2
tiktoken==0.5.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Streaming
sse-starlette==1.8.2
